--- index.css
@@ -1,9 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-#app {
-    position: relative; /* 確保子元素共享同一個層疊上下文 */
-    z-index: 0; /* 初始化層級 */
-  }
-
