+++ src/components/googleLogin.vue
@@ -53,23 +53,20 @@ const onLogin = (res) => {
   axios.post(`${API_URL}/auth/verify-token`, res, axiosOptions)
     .then((res) => {
       if (res.data.exists) {
-
+        console.log('用戶資料：', res.data.user) // 看看收到什麼資料
         // 用戶存在（可能是一般註冊或 Google 註冊），直接登入
         userData.value = res.data.user
         isLoggedIn.value = true
         localStorage.setItem(STORAGE_KEY, userData.value.userId) // 改存 UID
+        console.log('登入成功')
       } else {
-        // 用戶不存在，詢問是否要註冊，目前先用 sweetalert        
+        // 用戶不存在，詢問是否要註冊，這邊可以考慮搞個彈窗或是 sweetalert
         Swal.fire({
           title: '註冊確認',
-          text: `此 email 尚未註冊，是否要註冊新帳號？ \n ${res.data.googleData.email}`,
+          text: '此 email 尚未註冊，是否要註冊新帳號？',
           showCancelButton: true,
           confirmButtonText: '註冊',
-          cancelButtonText: '取消',
-          confirmButtonColor: '#000000', // 註冊按鈕設為黑色背景
-          customClass: {
-            confirmButton: 'swal2-confirm-custom' // 自定義class
-          }
+          cancelButtonText: '取消'
         }).then((result) => {
           if (result.isConfirmed) {
             handleRegister(res.data.googleData)
@@ -87,7 +84,8 @@ const handleLogout = () => {
   userData.value = null
   isLoggedIn.value = false
   localStorage.removeItem(STORAGE_KEY) // 移除 localStorage
-
+  console.log("登出成功");
+  
   // 等待 DOM 更新後再重新渲染按鈕
   setTimeout(() => {
     initializeGoogle()
@@ -116,7 +114,7 @@ const initializeGoogle = () => {
   // 渲染 Google 登入按鈕~~~ 可以選樣式啦
   window.google.accounts.id.renderButton(
     document.getElementById("googleButton"),
-    {
+    { 
       theme: "outline",  // outline 或 filled_blue
       size: "large",     // large 或 medium
       shape: "pill", // rectangular 或 pill
@@ -124,6 +122,7 @@ const initializeGoogle = () => {
       logo_alignment: "left"
     }
   )
+
 }
 
 onMounted(() => {
@@ -142,10 +141,3 @@ onMounted(() => {
   }
 })
 </script>
-
-<style scoped>
-.swal2-confirm-custom {
-  color: #ffffff !important;
-}
-</style>
-
