+++ src/views/ProductList.vue
@@ -1,18 +1,18 @@
 <script setup>
 import ProductItem from "@/components/ProductItem.vue";
 import { ref,computed,onMounted, watch } from "vue"
-import { storeToRefs } from "pinia";
 import { useRoute } from 'vue-router'
+import { storeToRefs } from "pinia";
 import { useProductStore } from '@/stores/products'
-import SmallCart4 from "@/components/SmallCart4.vue";
+import SmallCart from "@/components/SmallCart.vue";
 
 const route = useRoute()
-const productStore = useProductStore()
 const cartVisible = ref(false); // 控制 SmallCart 的顯示
-const { categoryTitle,productList, pageValue, sortValue, sortOptions, pageOptions, currentPage, itemsPerPage, paginatedProducts, paginationOnClickHandler  } = storeToRefs(productStore)
 const toggleCart = () => {
   cartVisible.value = !cartVisible.value;
 };
+const ProductStore = useProductStore()
+const { categoryTitle, productList, pageValue, sortValue, sortOptions, pageOptions, currentPage, itemsPerPage, paginatedProducts, paginationOnClickHandler } = storeToRefs(ProductStore)
 
 // 監聽路由參數變化
 watch(() => route.params.category, async (newCategory) => {
@@ -46,8 +46,6 @@ const cartItemCount = computed(() => {
   return cartItems.value.reduce((total, item) => total + item.quantity, 0);
 });
 
-
-
 </script>
 
 
@@ -56,14 +54,17 @@ const cartItemCount = computed(() => {
     <div class="headerContainer px-1 mb-2 md:flex items-center">
       <h1 class=" py-5 text-xl">{{ categoryTitle }}</h1>
       <!-- 排序 -->
-      <div class="flex selectContainer">
-        <div class="relative flex items-center flex-1 mr-3 pageSelectItem">
-          <i class="absolute text-gray-500 transform -translate-y-1/2 fa-solid fa-arrow-up-short-wide left-3 top-1/2"></i>
-          <el-select placement="bottom" :fallback-placements="['bottom-start']" v-model="sortValue" placeholder="商品排序" size="large" class="pl-10">
-            <el-option v-for="item in sortOptions" :key="item.value" :label="item.label" :value="item.value"/>
+      <div class="selectContainer flex">
+        <div class="pageSelectItem  flex items-center relative mr-3 flex-1">
+          <i
+            class="fa-solid fa-arrow-up-short-wide absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-500"></i>
+          <el-select placement="bottom" :fallback-placements="['bottom-start']" v-model="sortValue" placeholder="商品排序"
+            size="large" class="pl-10">
+            <el-option v-for="item in sortOptions" :key="item.value" :label="item.label" :value="item.value" />
           </el-select>
         </div>
         <!-- 每頁資料筆數 -->
+
         <div class="pageSelectItem  flex items-center relative flex-1">
           <i class="fa-solid fa-bars fa-rotate-90 absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-500"></i>
           <el-select placement="bottom" :fallback-placements="['bottom-start']" v-model="pageValue"
@@ -78,53 +79,14 @@ const cartItemCount = computed(() => {
 
     <!-- 產品列表 -->
     <div class="flex flex-wrap">
-      <ProductItem v-for="(item, index) in paginatedProducts" :key="item.id" :id="item.id" :title="item.title" :price="item.price"
-      :originalPrice="item.originalPrice" :frontImg="item.frontImg" :backImg="item.backImg" @addToCart="handleAddToCart"
-    @removeFromCart="removeFromCart"
-    @updateQuantity="updateQuantity"
-    class="md:col-6 lg:col-3"/>
-      <!-- 顯示 coBrandingProductList 商品 -->
-  <!-- <ProductItem
-    v-for="item in coBrandingProductList"
-    :key="item.id"
-    :id="item.id"
-    :title="item.title"
-    :price="item.price"
-    :originalPrice="item.originalPrice"
-    :frontImg="item.frontImg"
-    :backImg="item.backImg"
-    @addToCart="handleAddToCart"
-    @removeFromCart="removeFromCart"
-    @updateQuantity="updateQuantity"
-    class="md:col-6 lg:col-3"
-  /> -->
-</div>
-
-<div class="flex flex-wrap">
-  <!-- 顯示 productList 商品 -->
-  <!-- <ProductItem
-    v-for="item in productList"
-    :key="item.id"
-    :id="item.id"
-    :title="item.title"
-    :price="item.price"
-    :originalPrice="item.originalPrice"
-    :frontImg="item.frontImg"
-    :backImg="item.backImg"
-    @addToCart="handleAddToCart"
-    @removeFromCart="removeFromCart"
-    @updateQuantity="updateQuantity"
-    class="md:col-6 lg:col-3"
-  /> -->
-</div>
-
-<!-- 顯示小購物車，只顯示一次 -->
-<SmallCart4 :visible="cartVisible" :cartItems="productStore.cartItems" />
-
+      <ProductItem v-for="(item, index) in paginatedProducts" :key="item.id" :title="item.title" :price="item.price"
+        :orginalPrice="item.orginalPrice" :frontImg="item.frontImg" :backImg="item.backImg"  class="md:col-6 lg:col-3"
+        @addToCart="handleAddToCart" @removeFromCart="removeFromCart" @updateQuantity="updateQuantity" />
+    </div>
 
     <!-- 分頁 -->
-    <div class="flex justify-center md:relative md:mb-12">
-      <vue-awesome-paginate class="text-sm text-gray-500 md:absolute md:right-0" :total-items="productList.length"
+    <div class="flex justify-center md:relative  md:mb-12">
+      <vue-awesome-paginate class=" md:absolute md:right-0 text-gray-500 text-sm" :total-items="productList.length"
         :items-per-page="itemsPerPage" :max-pages-shown="5" v-model="currentPage" @click="paginationOnClickHandler"
         :hide-prev-next-when-ends="true" link-url="/products?page=[page]" />
     </div>
@@ -162,7 +124,7 @@ const cartItemCount = computed(() => {
 }
 
 
-:deep(.el-select__placeholder span){
+:deep(.el-select__placeholder span) {
   color: #000;
 }
 
@@ -178,12 +140,13 @@ const cartItemCount = computed(() => {
 }
 
 /* 去除 focused 時的邊框 */
-:deep(.el-select__wrapper.is-focused){
+:deep(.el-select__wrapper.is-focused) {
   box-shadow: 0 0 0 0px #fff inset;
 }
 
-:deep(.el-select__wrapper.is-hovering:not(.is-focused)){
+:deep(.el-select__wrapper.is-hovering:not(.is-focused)) {
   box-shadow: 0 0 0 0px #fff inset;
+
 }
 
 /* option 的 hover 樣式 */
@@ -194,15 +157,13 @@ const cartItemCount = computed(() => {
 }
 
 
-
-.el-select-dropdown__item.is-hovering{
+.el-select-dropdown__item.is-hovering {
   background-color: #000;
   color: #fff;
 }
 
 
-
-@media screen and (768px <= width) {
+@media screen and (768px <=width) {
   .pageSelectItem {
     width: 200px;
     height: 50px;
@@ -210,6 +171,6 @@ const cartItemCount = computed(() => {
 
   .el-select--large .el-select__wrapper {
     padding: 4px;
-}
+  }
 }
 </style>
